requirements.txt
discord
requests
groq
google
google.genai
google.generativeai
aiofiles
edge_tts
aiosqlite
pytz
pillow
python-dotenv
pandas
PyMuPDF
python-docx
openpyxl
bs4
googlesearch-python
undetected_chromedriver
setuptools